the support and its menu has been finished

+++ src/components/Modal/confirm/DeleteUserModal.tsx
@@ -3,7 +3,6 @@ import { useAppContext } from '../../../context/AppContext';
 import type { AppContextIn } from '../../../Interface/InApp';
 import { useAuth } from '../../../context/AuthContext';
 import type { AuthContextIn } from '../../../Interface/InAuth';
-import { useEffect } from 'react';
 
  interface ModalProp{
     id: number
@@ -15,6 +14,12 @@ function DeleteUserModal(prop: ModalProp) {
   const authContext = useAuth() as AuthContextIn
 
   async function deleteU(){
+
+    if(prop.id == authContext.user.id as number){
+      alert("Error, estas logeado con este usuario")
+      return
+    }
+
     const val = await authContext.deleteUser(prop.id)
     if(val){
       showConfirmModal(false)
@@ -31,6 +36,7 @@ function DeleteUserModal(prop: ModalProp) {
           onHide={() => {showConfirmModal(false)}} 
           message="Desea eliminar el producto seleccionado?" 
           header="Eliminar" icon="pi pi-exclamation-triangle" 
+          acceptLabel='Si'
     />
 </>
   )
